weather_fetcher_GCP_CF/src/requirements.txt
python-dotenv==0.21.0
requests==2.32.3